requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.1.0
SQLAlchemy==1.4.46
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.111.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
werkzeug==2.3.7
alembic